requirements.txt
numpy
pandas
matplotlib
tqdm
pyyaml
scikit-image
mpi4py
torch
pytorch-lightning
diffusers
datasets
opencv-python
nibabel==5.1.0
wandb==0.15.5
arrow==1.2.3
blessed==1.20.0
click==8.1.3
croniter==1.3.15
dateutils==0.6.12
deepdiff==6.3.0
fastapi==0.99.1
filelock==3.12.2
h11==0.14.0
inquirer==3.1.3
itsdangerous==2.1.2
lightning==2.0.4
lightning-cloud==0.5.37
lightning-utilities==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
ordered-set==4.1.0
pydantic==1.10.10
pyjwt==2.7.0
python-editor==1.0.4
python-multipart==0.0.6
readchar==4.0.5
rich==13.4.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
torchmetrics==0.11.4
torchvision==0.15.2
uvicorn==0.22.0
websockets==11.0.3
blobfile==2.0.2
Pillow==9.4.0
requests==2.29.0
black==22.6.0
flake8==6.0.0
isort==5.9.3
pre-commit==3.3.3
